ARM64: dts: qcom: vayu panels: Don't send unnecessary HSA/HSE packets
According to the MIPI DSI v1.3 specification, HSA (Horizontal Sync
Active) and HSE (Horizontal Sync End) packets that indicate the
beginning and end of a line should *not* be sent if the panel is
running in non-burst sync event mode or burst mode.

This also frees up some bandwidth on the DSI link and reduces data,
which should increase chances of lower-quality panels being able to
tolerate higher refresh rates.

Suggested-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: Danny Lin <danny@kdrag0n.dev>
Signed-off-by: Jebaitedneko <Jebaitedneko@gmail.com>
Change-Id: If42207118845fe6982767edb9655467dafd410a2
Signed-off-by: bagaskara815 <bagaskara815@gmail.com>